% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{todo}
\alias{todo}
\title{Add a todonotes to a Rmarkdown}
\usage{
todo(x = "", color = "orange")
}
\arguments{
\item{x}{text to be included in the todo.
This text is modified with rmarkdown render for latex output and with
custom code for html output.}

\item{color}{orange by default, this is the color background of the todonote.}
}
\description{
Add a todonotes to a Rmarkdown
}
\details{
For now, the package is based on the \code{todonotes} package in latex,
but only use the inlince option. Works will be done to include
\code{listoftodo} and a greater variety of todonotes.
}
\author{
Maxime Jaunatre
}
