% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{usefilter_lua}
\alias{usefilter_lua}
\title{Load a Lua filter}
\usage{
usefilter_lua(name)
}
\arguments{
\item{name}{The Lua filter name}
}
\description{
Load a Lua filter using R code.
}
\examples{
usefilter_lua("color-text.lua")
}
